requirements.txt
Flask==2.0.1
Pillow==8.4.0
requests==2.26.0
ultralytics==1.0.6
